feat: manage_kong = manage_service + register_app
